requirements.txt
requests
pyyaml
json